strata/requirements.txt
fastapi==0.110.0
uvicorn==0.27.1
httpx==0.27.2
PyYAML==6.0.2